Parse, render multi-paragraph footnotes.
Parsing: FnDef will greedily grab all following indented code.  As far as I can
tell, this conforms to the behavior of pandoc and multimarkdown.

Rendering: If the final block in the footnote is a paragraph, the "return to
article" link will render in that paragraph.  Otherwise, a new paragraph is
appended to the footnote containing the return link.  This mimics pandoc's
behavior -- multimarkdown will append it to the last paragraph of the footnote,
regardless of what follows.