Adds failing tests for loose and tight lists
This is related to issue #226.

The CommonMark spec says the following about lists:

> A list is loose if any of its constituent list items are separated by
> blank lines, or if any of its constituent list items directly contain
> two block-level elements with a blank line between them. Otherwise a
> list is tight. (The difference in HTML output is that paragraphs in a
> loose list are wrapped in <p> tags, while paragraphs in a tight list
> are not.)

In Earmark, many lists without blank line separators are rendered as if
they were loose. For example, given a nested list, the outer list is
always treated as loose even when there are no blank lines. This appears
to be an issue with any item that has multiple content blocks.

Another issue to consider with nested lists: we need to determine
whether or not a list is loose or tight at each level independently
based on their direct constituents (as described above). So it's
possible to have nested lists with many levels that each have varying
looseness/tightness.

This commit adds tests derived from the CommonMark spec to better cover
these edge cases. Some are currently failing because of the problems
above.