Add an optimized implementation for HTML escaping and part of smartypants
This patch replaces {String,Regex}.replace for HTML escaping and most
smartypants substitutions in favor of a faster custom implementation
that does only one pass using binary matches and outputs iodata.

The speedup in the benchmark is about 7-10% (29117.22 µs/op vs
27083.15 µs/op on a particular run). Note that it could be optimized
further by removing the remaining regexes, however the gained time
doesn't seem worth it for the added complexity.

This patch was written mostly as an experiment, feel free to reject
if you don't feel the added complexity is worth the speedup.